+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2021-05-17
+
 ### Added
 
 - Also delete legacy `companyd` organizations when deleting `Organization` CRs.
@@ -50,7 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release.
 
-[Unreleased]: https://github.com/giantswarm/organization-operator/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/giantswarm/organization-operator/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/giantswarm/organization-operator/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/giantswarm/organization-operator/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/giantswarm/organization-operator/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/giantswarm/organization-operator/compare/v0.2.0...v0.3.0